refactor prereqs
