+++ src/pages/index.js
@@ -1,6 +1,6 @@
 // CourseCatalog.js
 
-import React, { use, useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Inter } from 'next/font/google'
 const inter = Inter({ subsets: ['latin'] })
 import Card from "../components/card"
@@ -15,7 +15,6 @@ import {
   PopoverTrigger,
   PopoverContent,
   PopoverBody,
-  Stack,
   RangeSlider,
   RangeSliderTrack,
   RangeSliderFilledTrack,
@@ -54,6 +53,7 @@ const CourseCatalog = () => {
   const [searchTerm, setSearchTerm] = useState("");
   const [displayLanding, setDisplayLanding] = useState(true);
 
+  // Function to change from initial page to search result page
   function changeLanding(event) {
     if (event.length >= 2) {
       setDisplayLanding(!displayLanding);
@@ -71,7 +71,7 @@ const CourseCatalog = () => {
     return query;
   }
 
-
+  // On a filter change, rerun the search
   useEffect(() => {
     search();
   }, [JSON.stringify(selectedSubjects), JSON.stringify(selectedSemesters), JSON.stringify(selectedGenEds), transform(searchTerm), creditsMin, creditsMax, levelMin, levelMax, JSON.stringify(levels), JSON.stringify(scheds)]);
@@ -87,6 +87,7 @@ const CourseCatalog = () => {
   }, [displayLanding]);
 
 
+  // Search function, set courses to the result of the search
   const search = async (event) => {
     if (searchTerm.length <= 1 && selectedSubjects.length == 0 && selectedSemesters.length == 0 && selectedGenEds.length == 0) {
       setCourses([]);
@@ -114,6 +115,8 @@ const CourseCatalog = () => {
     }
   };
 
+
+  // scroll to top listener
   useEffect(() => {
     try {
       if (typeof document !== 'undefined') {
@@ -132,6 +135,8 @@ const CourseCatalog = () => {
       }
     } catch (e) { }
   }, []);
+
+
 
   return (
     <>
